moved uatraits
