+++ src/Core/UatraitsFast/cached.h
@@ -4,10 +4,10 @@
 
 #include <ext/unlock_guard.h>
 
-#include <Common/UatraitsFast/uatraits-fast.h>
+#include <Core/UatraitsFast/uatraits-fast.h>
 
 
-/** Вариант UATraits с кэшом готовых результатов и сохранением регистра в выходных полях.
+/** Вариант UATraits с кэшом готовых результатов.
   * В качестве кэша используется cache table
   *  (open addressing хэш-таблица фиксированного размера
   *   без механизма разрешения коллизий, с автоматическим вытеснением)
@@ -17,10 +17,10 @@
   * Mutex блокируется на короткое время, которое не включает в себя тяжёлые вычисления,
   *  поэтому масштабирование на несколько потоков нормальное.
   */
-class UATraitsCaseSafeCached : private boost::noncopyable
+class UATraitsCached : private boost::noncopyable
 {
 public:
-    UATraitsCaseSafeCached(const std::string & browser_path_, const std::string & profiles_path_, const std::string & extra_path_)
+    UATraitsCached(const std::string & browser_path_, const std::string & profiles_path_, const std::string & extra_path_)
         : table(std::make_unique<Cell []>(cache_size))
         , impl(browser_path_, profiles_path_, extra_path_)
     {
@@ -32,29 +32,25 @@ public:
             table = std::make_unique<Cell []>(cache_size);
     }
 
-    /** См. документацию к функции UATraits::detectCaseSafe().
+    /** См. документацию к функции UATraits::detect.
       */
-    void detectCaseSafe(
-        StringRef user_agent,
-        StringRef user_agent_lower,
-        StringRef profile,
-        StringRef x_operamini_phone_ua,
-        StringRef x_operamini_phone_ua_lower,
+    void detect(
+        StringRef user_agent_lower, StringRef profile, StringRef x_operamini_phone_ua_lower,
         UATraits::Result & result,
         UATraits::MatchedSubstrings & matched_substrings) const
     {
-        if (!user_agent.size && !profile.size && !x_operamini_phone_ua.size)
+        if (!user_agent_lower.size && !profile.size && !x_operamini_phone_ua_lower.size)
             return;
 
-        if (profile.size || x_operamini_phone_ua.size)
+        if (profile.size || x_operamini_phone_ua_lower.size)
         {
-            impl.detectCaseSafe(user_agent, user_agent_lower, profile, x_operamini_phone_ua, x_operamini_phone_ua_lower, result, matched_substrings);
+            impl.detect(user_agent_lower, profile, x_operamini_phone_ua_lower, result, matched_substrings);
             return;
         }
 
         UInt128 hash;
 
-        sipHash128(user_agent.data, user_agent.size, hash.ui128);
+        sipHash128(user_agent_lower.data, user_agent_lower.size, hash.ui128);
 
         size_t index = hash.lo % cache_size;
 
@@ -74,7 +70,7 @@ public:
 
             {
                 ext::unlock_guard<std::mutex> unlock(mutex);
-                impl.detectCaseSafe(user_agent, user_agent_lower, profile, x_operamini_phone_ua, x_operamini_phone_ua_lower, result, matched_substrings);
+                impl.detect(user_agent_lower, profile, x_operamini_phone_ua_lower, result, matched_substrings);
             }
 
             cell.first = hash;
@@ -92,8 +88,7 @@ private:
 
     union UInt128
     {
-        struct
-        {
+        struct {
             UInt64 lo;
             UInt64 hi;
         };
