+++ src/Core/UatraitsFast/tests/uatraits-conditions.cpp
