+++ src/Core/UatraitsFast/tests/uatraits-conditions/ya.make
