+++ src/Core/UatraitsFast/tests/uatraits-fast-cached/ya.make
