+++ src/Core/UatraitsFast/tests/uatraits-fast-mt/ya.make
