+++ src/Core/UatraitsFast/tests/uatraits-fast-perf/ya.make
