+++ src/Core/UatraitsFast/tests/uatraits-fast-test/ya.make
