+++ src/Core/UatraitsFast/tests/uatraits-name-by-model/ya.make
