+++ src/Core/UatraitsFast/tests/uatraits-user-agent/ya.make
