+++ src/Core/UatraitsFast/tests/uatraits_mobile_phone_match/ya.make
