+++ src/Core/UatraitsFast/types/Browser.h
