+++ src/Core/UatraitsFast/types/OperatingSystem.cpp
