+++ src/Core/UatraitsFast/types/OperatingSystem.h
